daemon: Create kube-proxy replacement module for helpers
And move all related helpers from cmd/daemon_main.go into
cmd/kube_proxy_replacement.go to not bloat over the former.
